backend/requirements.txt
django
djangorestframework
psycopg2-binary
openai
googlemaps
django-cors-headers
python-dotenv
